requirements.txt
pandas
yfinance
schedule
winotify
pytz
python-dotenv
streamlit
plotly
